+++ app/test/e2e/task-get.spec.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-unused-vars,no-undef */
 const nock = require('nock');
 const chai = require('chai');
+const config = require('config');
 const Task = require('models/task.model');
 const { task } = require('rw-doc-importer-messages');
 const appConstants = require('app.constants');
@@ -122,7 +123,7 @@ describe('Task get tests', () => {
             }];
         await fakeTask.save();
 
-        nock(process.env.ELASTIC_URL)
+        nock(config.get('elasticsearch.host'))
             .get(`/_tasks/${encodeURIComponent(fakeTask.logs[0].elasticTaskId)}`)
             .reply(200, elasticTaskResponseObject);
 
@@ -177,7 +178,7 @@ describe('Task get tests', () => {
             }];
         await fakeTask.save();
 
-        nock(process.env.ELASTIC_URL)
+        nock(config.get('elasticsearch.host'))
             .get(`/_tasks/${encodeURIComponent(fakeTask.logs[0].elasticTaskId)}`)
             .reply(404, elasticTaskResponseObject);
 
@@ -232,7 +233,7 @@ describe('Task get tests', () => {
             }];
         await fakeTask.save();
 
-        nock(process.env.ELASTIC_URL)
+        nock(config.get('elasticsearch.host'))
             .get(`/_tasks/${encodeURIComponent(fakeTask.logs[0].elasticTaskId)}`)
             .reply(400, elasticTaskResponseObject);